Try to fix tidy on the shared ptr file
