+++ src/include/duckdb/common/weak_ptr_ipp.hpp
@@ -1,3 +1,7 @@
+#pragma once
+
+#include "duckdb/common/compatible_with_ipp.hpp"
+
 namespace duckdb {
 
 template <typename T, bool SAFE>
@@ -113,5 +117,8 @@ public:
 		return internal >= other.internal;
 	}
 };
+
+template <typename T>
+using unsafe_weak_ptr = weak_ptr<T, false>;
 
 } // namespace duckdb
